mcp_server/requirements.txt
websockets==10.4
python-dotenv==1.0.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.97.0 